backend/requirements.txt
fastapi==0.114.1
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.1.0
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-multipart==0.0.9
alembic==1.13.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==3.2.2
pytest==8.3.3
ruff==0.7.3
black==24.10.0
httpx==0.24.1
